nfs: Add richacl support
Add support for the "system.richacl" xattr in nfs.  The existing
"system.nfs4_acl" xattr on nfs doesn't map user and group names to uids
and gids; the "system.richacl" xattr does, and only keeps the
on-the-wire names when there is no mapping.  This allows to copy
permissions across different file systems.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>

BACKPORT Conflicts:
	fs/nfs/nfs4proc.c
	fs/nfs/nfs4xdr.c
	include/linux/nfs_xdr.h

Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>