nfsd: don't embed the fsnotify_mark inside of nfsd_file
The way that the fsnotify_mark is currently embedded inside the
nfsd_file is not safe. The nfsd_file is freed via call_rcu after the RCU
grace period settles, but fsnotify_marks are freed via a kthread after
the SRCU grace period has settled. It's possible that the nfsd_file
could be reallocated before the SRCU grace period has settled. If that
happens, then we'll end up with memory corruption as
fsnotify_mark_destroy will attempt to operate on the mark even though it
has been freed and possibly reallocated.

Fix this by not embedding the mark inside the nfsd_file and allocating
it separately. While we're at it, we can also stop allowing duplicate
marks by having it be a refcounted object. When we add a new nfsd_file,
we first attempt to see if there is already a mark on the inode. If so,
we take a reference to it otherwise we allocate a new mark object and
attach it to the inode. When the last reference is dropped, we detach it
from the inode and free it.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>