nfsd: add a NFSD_FILE_REFERENCED flag and let LRU code handle rotation
As Dave Chinner pointed out, we really don't need to rotate each
nfsd_file to the end of the LRU. If we simply keep track of a REFERENCED
flag on the object, then we can use that to tell whether there has been
any activity on the object. On the second pass through the shrinker
without that flag being reset, it'll be removed.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>