nfs: add "local io" support
Add client support for bypassing NFS for localhost reads, writes, and commits.

This is only useful when the client and the server are running on the same
host and in the same container.

This has dynamic binding with the nfsd module. Local i/o will only work if
nfsd is already loaded.

Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>