nfsd: Ensure files don't remain open forever in the filecache
We neither want nor need files to remain open forever in the filecache.
It is sufficient to keep them open for long enough that the cost of
a new open() is amortised.
This implementation just reuses the existing mark+sweep code from the
memory shrinker, triggering it every 60 seconds from a workqueue.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>