nfsd: Fix a livelock in the lru shrink/walk code
We do not want to use LRU_ROTATE or LRU_REMOVED_RETRY in a situation
where we might risk getting called by list_lru_walk().

The reason LRU_ROTATE is bad is that it moves the entry to the end of
the lru list. If it isn't freeable, then that means we livelock since
__list_lru_walk_one() uses list_for_each_safe() to traverse the list.
Furthermore, rotating the list in this situation means we clear
NFSD_FILE_REFERENCED, then retest the same entry.

The problem with LRU_REMOVED_RETRY is that it restarts the list
traversal, meaning that we again retest all entries that have
been skipped after having NFSD_FILE_REFERENCED cleared.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>