nfsd: fix bad fsnotify_mark handling in filecache code
fsnotify_mark_add will have hashed the mark passed in on success and
failure, so we can't just free one that has been allocated. We must
instead put the reference to the mark if the add fails.

This might be the source of the memory scribble that has been reported
in PD-11743, but we won't know for sure until it's merged and tested.

JIRA: PD-11743
Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>