nfs: the local_file_open operations should take a reference to the filp
In the non-pnfs case, the local_file doesn't change so this is not
mandatory, but with flexfiles we can end up with mirrors that can
come and go. When that happens we want to close the local_file and
(possibly) reopen it. Prepare for that by having the local_file_open
operation return a reference to the file, and have the callers put
that reference after they've used it.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>