nfs4: allow ff_local_open_fh to actually open the file
For some odd reason, it just goes and fetches the pointer out of the ds
currently. Allow it to attempt to open the file if the pointer is NULL.
While we're in here, go ahead and make it access the pointer under the
rcu_read_lock, since we'll need that later. With this we can drop
ff_layout_update_mirror_filp, as it'll be done dynamically as needed.

Note that we also "preload" the local_file in the flexfiles codepaths.
This is so that we can try to make an intelligent decision about whether
to attempt localio or not. More cleanup in this area to follow.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>