nfs: don't allow fake svc_rqst objects to be deferred
These requests are faked up in order to handle a localio request and
hence can't be placed on the deferred queue like a "normal" svc_rqst.

Set the thread_wait value to the same value that normal nfsd threads use,
and clear out the defer pointer and RQ_USEDEFERRAL flags to ensure that
they don't end up on the deferral queue.

JIRA: PD-18101
Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>