sunrpc: Add AUTH_NAME support
AUTH_NAME encodes all principals as string in "principal@domain" form.
It also supports a large number of principals - way more than the 18 in
an AUTH_UNIX cred.

Because the definition of an AUTH_NAME cred can be very large, we don't
want to send it with every RPC. Instead we follow the design of RPCSEC_GSS
and init / destroy a session encoded in the payload of a NULL ping
(in whatever rpc program it's being used for). Subsequent calls will
use the session id established from the init call to identify the cred
in use. The server may at any time respond with REJECTEDCRED which causes
the client to reestablish the session.

Signed-off-by: Weston Andros Adamson <dros@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>
Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>