auth_name: fix races in session state management
Allocate a ref counted "name_session" object which is set atomically
using an rcu protected pointer.

This changes the refresh path to block all callers to wait on the init
call, but specifically allows NULL procedures through because the init
call (along with destroy) uses the same credential.

This fixes PD-27944

Signed-off-by: Weston Andros Adamson <dros@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>