Add mkdirat2() syscall to allow passing of a flag argument
Add a syscall mkdir2() and an internal VFS inode_operation mkdir2() to allow
passing of a flag argument.
Allow Samba to specify that it wants the created directory to inherit the
unmodified ACL from its parent.

Note: This is a hack to tide us over until we have proper support in the
upstream kernel for this kind of system call.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>
Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>
Signed-off-by: Mike Snitzer <snitzer@hammerspace.com>