Ensure that we also drop the ref taken on the mark when we created it.
When we initialize a mark we get a ref on it. Then when we add the mark
to an inode we get another ref on the mark. Thus, when we clean up, we
have to delete the mark (which drops one ref) and then put the mark for
its ref count to go all the way to zero and for it to be freed.

Signed-off-by: Richard Sharpe <richard.sharpe@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>