NFSv4: Fix up settings for setting ARCHIVE/BACKUP_TIME
If the user tries to set ARCHIVE, but the server doesn't have support
for that bit, then we try to manipulate time_backup instead.
If the user sets NFS_FA_FLAG_ARCHIVE, then we set time_backup to the
current inode mtime. Otherwise we set time_backup to the file create time.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Signed-off-by: Lance Shelton <lance.shelton@hammerspace.com>