pNFS/flexfiles: Don't turn off localio on open errors
There are lots of reasons why an open() could fail on the DA-DS local
filesystem. Everything from fencing to upcall timeouts will cause
errors to be returned. However the correct way to deal with those
errors is to report them to the server, not to unilaterally turn off
the DA-DS functionality.

Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>